Remove display project title to the dom.js file
