+++ src/project.js
@@ -1,3 +1,4 @@
+import { displayProjectTitle } from './dom'; // eslint-disable-line
 import { task, showTask } from './task'; // eslint-disable-line
 
 
@@ -63,24 +64,7 @@ function myEvent() {
   }
 }
 
-const displayProjectTitle = () => {
-  let i = 0;
-  document.querySelector('#project-container').innerHTML = '';
-  projectsArray.forEach(project => {
-    const projectDiv = document.createElement('div');
-    projectDiv.setAttribute('class', 'color-white px-4 bolden show-project mx-4 pt-3 mb-2 project-div');
-    projectDiv.setAttribute('id', i);
-    const titleparagraph = document.createElement('p');
-    const titleText = document.createTextNode(`${project.title}`);
-    titleparagraph.appendChild(titleText);
-    projectDiv.appendChild(titleparagraph);
-
-    const projectContainer = document.getElementById('project-container');
-    projectContainer.appendChild(projectDiv);
-    i += 1;
-  });
-  myEvent();
-};
+displayProjectTitle();
 
 
 const submit = () => {
@@ -108,5 +92,5 @@ const newProject = () => {
 displayProjectTitle();
 
 export {
-  projectCreation, newProject, saveArray, projectsArray,
+  projectCreation, newProject, saveArray, projectsArray, myEvent,
 };
