+++ src/task.js
@@ -1,5 +1,5 @@
 import { projectsArray, saveArray } from './project'; // eslint-disable-line
-import { myTask } from './dom'; // eslint-disable-line
+import { myShowTask, myTask } from './dom'; // eslint-disable-line
 
 const taskCreation = (description, dueDate, priority, project) => {
   project.todos.push({
@@ -19,62 +19,15 @@ newTaskForm.style.display = 'none';
 
 
 const showTask = (project) => {
-  myTask(project);
+  myShowTask(project);
 };
 
 
 const task = (projectId) => {
-  document.querySelector('#task-container').innerHTML = '';
-  const project = projectsArray[projectId];
-  const taskTitleParagraph = document.createElement('p');
-  const taskTextTitle = document.createTextNode(`Project: ${project.title}`);
-  taskTitleParagraph.appendChild(taskTextTitle);
-
-  const taskDescriptionParagraph = document.createElement('p');
-  const taskDescriptionText = document.createTextNode(`Description: ${project.description}`);
-  taskDescriptionParagraph.appendChild(taskDescriptionText);
-
-  const taskDiv = document.createElement('div');
-  taskDiv.setAttribute('class', 'color-white px-4 pt-5 bold-text');
-
-
-  taskDiv.appendChild(taskTitleParagraph);
-  taskDiv.appendChild(taskDescriptionParagraph);
-
-
-  const taskContainer = document.getElementById('task-container');
-  taskContainer.appendChild(taskDiv);
-
-  // Add task button
-  const newTaskButton = document.createElement('button');
-  newTaskButton.setAttribute('class', 'btn btn-success px-5 mx-4 my-3');
-  newTaskButton.setAttribute('id', 'new-task');
-  const newTaskText = document.createTextNode('Add A New "To-Do"');
-  newTaskButton.appendChild(newTaskText);
-  taskContainer.appendChild(newTaskButton);
-
-  // display new task form
-  const displayTaskForm = document.getElementById('new-task');
-  displayTaskForm.onclick = () => {
-    newTaskForm.style.display = 'block';
-  };
-
-  const submitTaskButton = document.getElementById('submit-task');
-
-  submitTaskButton.onclick = () => {
-    // retrieve value from form input
-    const descriptionTask = document.getElementById('exampleInputDescription').value;
-    const dueDateTask = document.getElementById('exampleInputDueDate').value;
-    let priority;
-
-
-    taskCreation(descriptionTask, dueDateTask, priority, project);
-    newTaskForm.style.display = 'none';
-
-    showTask(project);
-    saveArray();
-  };
+  myTask(projectId);
 };
 
 
-export { task, showTask };
+export {
+  task, showTask, taskCreation, newTaskForm,
+};
